Update LocateButton to set user location
